Nya reglering i korridor och korsning

+++ WISH_Gait_Engine.h
@@ -25,15 +25,15 @@ volatile float Y_Step_Length; // mm
 volatile int Angular_Step_Length; // grader*10 i heltal
 
 // Temporära parametrar från bussen
-volatile float X_Speed; // mm
+ volatile float X_Speed; // mm
 // volatile float X_Step_Length_temp; // mm
 // volatile float Y_Step_Length_temp; // mm
 // volatile int Angular_Step_Length_temp; // grader i heltal
 
 // Vinklar för kroppens rotation
-volatile float Y_Pitch; // grader
-volatile float X_Roll; // grader
-volatile float Z_Yaw; // grader
+volatile int Y_Pitch; // grader*10
+volatile int X_Roll; // grader*10
+volatile int Z_Yaw; // grader*10
 
 // Tabellerad SIN och COS värden
 volatile float SIN_TABLE[401];
@@ -50,7 +50,7 @@ void Emergency_Stop();
 
 void set_frame(leg_info* leg, float x_local, float y_local); // Beräkna benets position under en del-rörelse
 void tripod_gait(); // Sekvensiera benrörelse
-void body_rotate(leg_info* leg, float x_angle, float y_angle, float z_angle); // Rotera kroppen
+void body_rotate(leg_info* leg, int x_angle, int y_angle, int z_angle); // Rotera kroppen
 
 void tripod_climb_gait(float z_height);
 void set_frame_climb(leg_info* leg, float x_local, float y_local, float z_height);
